+++ spring-cloud-tencent-polaris-loadbalancer/src/main/java/com/tencent/cloud/polaris/loadbalancer/config/PolarisLoadBalancerAutoConfiguration.java
@@ -24,7 +24,6 @@ import com.tencent.polaris.factory.api.RouterAPIFactory;
 import com.tencent.polaris.router.api.core.RouterAPI;
 
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.netflix.ribbon.RibbonAutoConfiguration;
@@ -46,15 +45,12 @@ import org.springframework.context.annotation.Configuration;
 public class PolarisLoadBalancerAutoConfiguration {
 
 	@Bean
-	@ConditionalOnMissingBean
 	public PolarisLoadBalancerProperties polarisLoadBalancerProperties() {
 		return new PolarisLoadBalancerProperties();
 	}
 
-	@Bean(name = "polarisRoute")
-	@ConditionalOnMissingBean
-	public RouterAPI polarisRouter(SDKContext polarisContext) throws PolarisException {
+	@Bean
+	public RouterAPI routerAPI(SDKContext polarisContext) throws PolarisException {
 		return RouterAPIFactory.createRouterAPIByContext(polarisContext);
 	}
-
 }