+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 ---
 
-- [Feature: Support parse ratelimit rule expression labels.](https://github.com/Tencent/spring-cloud-tencent/pull/183)
 
+- [Feature: Support parse ratelimit rule expression labels.](https://github.com/Tencent/spring-cloud-tencent/pull/183)
+- [fix:Router support request label.](https://github.com/Tencent/spring-cloud-tencent/pull/165)